requirements.txt
streamlit==1.52.2
google-api-python-client==2.187.0
pandas==2.3.3
python-dotenv==1.0.1
